+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger.php
@@ -5,6 +5,7 @@ use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\FooMessage;
 
 $container->loadFromExtension('framework', [
     'messenger' => [
+        'reset_on_message' =>  true,
         'routing' => [
             FooMessage::class => ['sender.bar', 'sender.biz'],
             BarMessage::class => 'sender.foo',