+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_multiple_failure_transports_global.php
@@ -3,6 +3,7 @@
 $container->loadFromExtension('framework', [
     'messenger' => [
         'failure_transport' => 'failure_transport_global',
+        'reset_on_message' =>  true,
         'transports' => [
             'transport_1' => [
                 'dsn' => 'null://',