+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_without_reset_on_message_legacy.php
@@ -1,0 +1,18 @@
+<?php
+
+use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\BarMessage;
+use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\FooMessage;
+
+$container->loadFromExtension('framework', [
+    'messenger' => [
+        'routing' => [
+            FooMessage::class => ['sender.bar', 'sender.biz'],
+            BarMessage::class => 'sender.foo',
+        ],
+        'transports' => [
+            'sender.biz' => 'null://',
+            'sender.bar' => 'null://',
+            'sender.foo' => 'null://',
+        ],
+    ],
+]);